# plugins/orchestrator-leaderboard/docs/openapi.yaml
openapi: 3.1.0
info:
  title: NaaP Orchestrator Leaderboard API
  version: 1.0.0
  description: |
    Public REST API exposed by the **orchestrator-leaderboard** plugin.

    The API lets external SDKs and gateways:

    1. **Build a discovery plan** — declare a reusable orchestrator selection
       policy (capabilities + filters + SLA weights + sort).
    2. **Push the plan** — `POST /plans` persists it server-side and returns a
       stable `id`.
    3. **Get plan-executed results** — `GET /plans/{id}/results` returns the
       latest ranked orchestrators per capability, lazily evaluated and cached.

    For ad-hoc, stateless ranking (no persisted plan) use `POST /rank`.

    Source of truth:
      - Routes: `apps/web-next/src/app/api/v1/orchestrator-leaderboard/**`
      - Schemas: `apps/web-next/src/lib/orchestrator-leaderboard/types.ts`
      - Persistence: `prisma model DiscoveryPlan` (`plugin_orchestrator_leaderboard` schema)
  contact:
    name: NaaP Team
    email: team@naap.io
    url: https://naap.io
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://{host}
    description: Self-hosted NaaP deployment
    variables:
      host:
        default: app.naap.io
        description: NaaP web-next host (no trailing slash)
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerJwt: []
  - bearerApiKey: []
paths:
...
  /api/v1/orchestrator-leaderboard/plans/{id}:
    parameters:
      - $ref: "#/components/parameters/PlanId"
    get:
      tags: [Discovery Plans]
      summary: Get a discovery plan definition
      operationId: getPlan
      parameters:
        - $ref: "#/components/parameters/BillingProviderSlugQuery"
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          plan: { $ref: "#/components/schemas/DiscoveryPlan" }
                        required: [plan]
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Discovery Plans]
      summary: Update a discovery plan
      operationId: updatePlan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePlanInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          plan: { $ref: "#/components/schemas/DiscoveryPlan" }
                        required: [plan]
        "400":
          description: Validation failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Discovery Plans]
      summary: Delete a discovery plan
      operationId: deletePlan
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: true }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/orchestrator-leaderboard/plans/{id}/results:
    parameters:
      - $ref: "#/components/parameters/PlanId"
    get:
      tags: [Discovery Plans]
      summary: Get the latest evaluated results for a plan
      description: |
        Lazy evaluation:

        - If a fresh evaluation is cached, returns immediately (`X-Cache-Age`
          will be > 0).
        - Otherwise the plan is evaluated against the latest global dataset,
          cached, and returned.

        Disabled plans return `400`.
      operationId: getPlanResults
      parameters:
        - $ref: "#/components/parameters/BillingProviderSlugQuery"
      responses:
        "200":
          description: Ok
          headers:
            Cache-Control:
              schema: { type: string, example: "private, max-age=10" }
            X-Cache-Age:
              schema: { type: integer, description: "Cache age in milliseconds" }
            X-Refresh-Interval:
              schema: { type: integer, description: "Configured refresh interval in milliseconds" }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          data: { $ref: "#/components/schemas/PlanResults" }
                        required: [data]
        "400":
          description: Plan disabled or evaluation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/orchestrator-leaderboard/plans/{id}/python-gateway:
    parameters:
      - $ref: "#/components/parameters/PlanId"
    get:
      tags: [Discovery Plans, Python Gateway Discovery]
      summary: Plan-scoped python-gateway discovery
      description: |
        Returns a bare JSON array of orchestrator addresses for the plan's
        capabilities, de-duplicated and tier-shuffled. Not wrapped in
        `{ success, data }`.
      operationId: getPlanPythonGateway
      parameters:
        - $ref: "#/components/parameters/BillingProviderSlugQuery"
      responses:
        "200":
          description: Ranked orchestrator addresses
          headers:
            Cache-Control:
              schema: { type: string, example: "private, no-store, must-revalidate" }
            X-Cache-Age:
              schema: { type: integer }
            X-Refresh-Interval:
              schema: { type: integer }
            X-Pymthouse-Manifest:
              schema: { type: string, enum: [empty] }
              description: Present when manifest filtering removed all capabilities.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PythonGatewayAddress" }
        "400":
          description: Invalid query, disabled plan, or evaluation failed
          content:
            text/plain:
              schema: { type: string }
        "401":
          description: Unauthorized
          content:
            text/plain:
              schema: { type: string }
        "404":
          description: Plan not found
          content:
            text/plain:
              schema: { type: string }
        "500":
          description: Evaluation failed
          content:
            text/plain:
              schema: { type: string }
...
  /api/v1/orchestrator-leaderboard/dataset/refresh:
    post:
      tags: [Admin]
      summary: Trigger a global dataset refresh
      description: |
        Two auth paths:

        1. **CRON_SECRET** — Vercel cron (hourly). Honours the configured
           refresh interval and skips if the dataset is still fresh.
        2. **User JWT with `system:admin`** — manual *Refresh Now* from the
           admin UI; bypasses the time gate.
      operationId: refreshDataset
      security:
        - bearerJwt: []
        - bearerCronSecret: []
      responses:
        "200":
          description: Refresh executed or skipped
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - type: object
                            properties:
                              refreshed: { type: boolean }
                              capabilities: { type: integer }
                              orchestrators: { type: integer }
                          - type: object
                            properties:
                              skipped: { type: boolean, const: true }
                              reason: { type: string }
                              nextRefreshInMs: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500":
          description: Refresh failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
components:
  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        NaaP user session JWT. Obtained from the NaaP login flow and stored
        in the `naap_auth_token` localStorage key on the web client. Plans
        are scoped to the caller's `teamId` and `ownerUserId`.
    bearerApiKey:
      type: http
      scheme: bearer
      description: |
        Service Gateway API key in the form `gw_xxxxxxxx`. Created from
        **Service Gateway → API Keys** in the NaaP dashboard. Recommended
        for external SDK and signer integrations.
    bearerCronSecret:
      type: http
      scheme: bearer
      description: |
        Internal `CRON_SECRET` shared with Vercel Cron. Not for end users.
  parameters:
    PlanId:
      name: id
      in: path
      required: true
      schema: { type: string, description: "Plan UUID/CUID returned by createPlan" }
    BillingProviderSlugQuery:
      name: billingProviderSlug
      in: query
      required: false
      schema: { $ref: "#/components/schemas/BillingProviderSlug" }
      description: |
        Filter or scope by billing provider. When present but empty, returns 400.
        When omitted, no provider filter is applied on list/get routes.
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    BadRequest:
      description: Invalid request parameters or validation failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Authenticated but lacking required permission
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: Resource not found or not visible to the caller
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: {}
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - UPSTREAM_ERROR
                - GATEWAY_TIMEOUT
                - INTERNAL_ERROR
            message: { type: string }
    Capability:
      type: string
      pattern: "^[a-zA-Z0-9_-]+$"
      maxLength: 128
      description: "Capability identifier (e.g. `streamdiffusion-sdxl`, `noop`)."
      examples: [streamdiffusion-sdxl, image-to-video, noop]
    LeaderboardFilters:
      type: object
      additionalProperties: false
      properties:
        gpuRamGbMin: { type: number, minimum: 0 }
        gpuRamGbMax: { type: number, minimum: 0 }
        priceMax: { type: number, minimum: 0 }
        maxAvgLatencyMs: { type: number, minimum: 0 }
        maxSwapRatio: { type: number, minimum: 0, maximum: 1 }
    SLAWeights:
      type: object
      additionalProperties: false
      description: |
        Weights for the SLA composite score. Each weight is in `[0, 1]`. Weights
        do not need to sum to 1 — they are normalised internally. Higher weight
        on a metric means that metric counts more (lower latency / swap / price
        is always better).
      properties:
        latency: { type: number, minimum: 0, maximum: 1, default: 0.4 }
        swapRate: { type: number, minimum: 0, maximum: 1, default: 0.3 }
        price: { type: number, minimum: 0, maximum: 1, default: 0.3 }
    PlanSortBy:
      type: string
      enum: [slaScore, latency, price, swapRate, avail]
    BillingProviderSlug:
      type: string
      enum: [daydream]
      description: Billing provider that scopes manifest-based capability filtering.
    PythonGatewayAddress:
      type: object
      required: [address]
      additionalProperties: false
      properties:
        address:
          type: string
          format: uri
          description: Reachable orchestrator base URL.
...
    UpdatePlanInput:
      type: object
      description: |
        Partial update — any subset of `CreatePlanInput` (excluding
        `billingPlanId`, which is immutable). Also supports `enabled`.
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        description: { type: string, maxLength: 1000 }
        capabilities:
          type: array
          minItems: 1
          maxItems: 50
          items: { $ref: "#/components/schemas/Capability" }
        topN: { type: integer, minimum: 1, maximum: 1000 }
        slaWeights:
          oneOf:
            - $ref: "#/components/schemas/SLAWeights"
            - type: "null"
        slaMinScore:
          oneOf:
            - { type: number, minimum: 0, maximum: 1 }
            - type: "null"
        sortBy:
          oneOf:
            - $ref: "#/components/schemas/PlanSortBy"
            - type: "null"
        filters:
          oneOf:
            - $ref: "#/components/schemas/LeaderboardFilters"
            - type: "null"
        enabled: { type: boolean }
        billingProviderSlug: { $ref: "#/components/schemas/BillingProviderSlug" }
    DiscoveryPlan:
      type: object
      required:
        - id
        - billingPlanId
        - name
        - capabilities
        - topN
        - enabled
        - createdAt
        - updatedAt
      properties:
        id: { type: string, description: "Plan UUID/CUID" }
        billingPlanId: { type: string }
        billingProviderSlug:
          oneOf:
            - $ref: "#/components/schemas/BillingProviderSlug"
            - type: "null"
        name: { type: string }
        description: { type: [string, "null"] }
        teamId: { type: [string, "null"] }
        ownerUserId: { type: [string, "null"] }
        capabilities:
          type: array
          items: { $ref: "#/components/schemas/Capability" }
        topN: { type: integer }
        slaWeights:
          oneOf:
            - $ref: "#/components/schemas/SLAWeights"
            - type: "null"
        slaMinScore: { type: [number, "null"] }
        sortBy:
          oneOf:
            - $ref: "#/components/schemas/PlanSortBy"
            - type: "null"
        filters:
          oneOf:
            - $ref: "#/components/schemas/LeaderboardFilters"
            - type: "null"
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    OrchestratorRow:
      type: object
      required:
        - orchUri
        - gpuName
        - gpuGb
        - avail
        - totalCap
        - pricePerUnit
      properties:
        orchUri:
          type: string
          format: uri
          description: Reachable orchestrator base URL.
        gpuName: { type: string, description: "GPU model name." }
        gpuGb: { type: number, description: "GPU VRAM in GB." }
        avail: { type: integer, description: "Currently available capacity." }
        totalCap: { type: integer, description: "Total advertised capacity." }
        pricePerUnit: { type: number, description: "Price per unit (gateway-defined)." }
        bestLatMs: { type: [number, "null"], description: "Best observed latency, ms." }
        avgLatMs: { type: [number, "null"], description: "Rolling average latency, ms." }
        swapRatio: { type: [number, "null"], description: "Swap ratio in [0, 1] — lower is better." }
        avgAvail: { type: [number, "null"], description: "Average availability over the window." }
        slaScore:
          type: number
          minimum: 0
          maximum: 1
          description: |
            Composite SLA score. Present only when the request supplied
            `slaWeights` (rank endpoint) or the plan has `slaWeights` set.
    PlanResults:
      type: object
      required: [planId, refreshedAt, capabilities, meta]
      properties:
        planId: { type: string }
        refreshedAt: { type: string, format: date-time }
        capabilities:
          type: object
          description: |
            Map of `capabilityName` → ranked rows. Each capability list is
            already sorted, filtered, and capped to the plan's `topN`.
          additionalProperties:
            type: array
            items: { $ref: "#/components/schemas/OrchestratorRow" }
        plan:
          type: object
          description: Plan metadata echoed for convenience.
          properties:
            name: { type: string }
            description: { type: [string, "null"] }
            capabilities:
              type: array
              items: { $ref: "#/components/schemas/Capability" }
            topN: { type: integer }
        meta:
          type: object
          required: [totalOrchestrators, refreshIntervalMs, cacheAgeMs]
          properties:
            totalOrchestrators: { type: integer }
            refreshIntervalMs: { type: integer }
            cacheAgeMs: { type: integer }
...